requirements.txt
Flask==2.0.1 
flask-cors==3.0.10 
mysql-connector-python==8.0.29 
py-solc-x==1.1.1 
web3==6.0.0 
python-dotenv==0.19.0 
gunicorn==20.1.0